requirements.txt
torch~=1.9.1+cpu
numpy~=1.20.3
scikit-video~=1.1.11
scipy~=1.7.1
torchvision~=0.10.1+cpu
scikit-image~=0.18.3
Pillow~=8.3.2
pydub~=0.25.1
sda~=0.2
setuptools~=57.0.0